move logging to central location
