+++ src/cameras/wincamd.py
@@ -343,31 +343,6 @@ class WinCamD(cam.Camera):
 		self.apertureOpen = False
 		
 		return ret
-	
-	def log(self, msg: str, loglevel: int = logging.INFO):
-		"""Handles the logging to easily switch between different ways of handling
-
-		Parameters
-		----------
-		msg : str
-			The log message
-		loglevel : int
-			enum in https://docs.python.org/3/library/logging.html#logging-levels,
-			see https://github.com/python/cpython/blob/d730719b094cb006711b1cd546927b863c173b31/Lib/logging/__init__.py
-
-			CRITICAL = 50
-			FATAL = CRITICAL
-			ERROR = 40
-			WARNING = 30
-			WARN = WARNING
-			INFO = 20
-			DEBUG = 10
-			NOTSET = 0
-		"""
-
-		logging.log(loglevel, msg)
-		if loglevel >= logging.DEBUG:
-			print(f"{logging.getLevelName(loglevel)}: {msg}")
 
 	def __exit__(self, e_type, e_val, traceback):
 		if self.apertureOpen:
